feat(hooks): rename onDispose decorator to onContainerDisposed
Clarifies that the hook fires when the container/scope is disposed,
mirroring the onConstruct naming. Updates lib/hooks/onDispose.ts (now
onContainerDisposed.ts), the onDisposeHooksRunner export, tests, and
docs. addOnDisposeHook/AddOnDisposeHookModule/OnDisposeHook are
unchanged since they describe the lower-level container dispose-hook
registration mechanism, not the decorator itself.

BREAKING CHANGE: the `onDispose` decorator export is renamed to
`onContainerDisposed`. Update imports and `@onDispose(...)` usages to
`onContainerDisposed`/`@onContainerDisposed(...)`.
