+++ .readme.hbs.md
@@ -19,7 +19,7 @@ provider pipelines, aliases, and custom injector strategies.
 - clean API for classes, keys, tokens, aliases, and scopes
 - no global container object; pass containers and scopes explicitly
 - supports tagged application, request, transaction, page, and widget scopes
-- decorator support with `@register`, `@inject`, `@onConstruct`, and `@onDispose`
+- decorator support with `@register`, `@inject`, `@onConstruct`, and `@onContainerDisposed`
 - can [inject properties](#inject-property)
 - can inject [lazy dependencies](#lazy)
 - composable provider and registration pipelines
@@ -53,7 +53,7 @@ provider pipelines, aliases, and custom injector strategies.
 - [Module](#module)
 - [Hook](#hook) `@hook`
   - [OnConstruct](#onconstruct) `@onConstruct`
-  - [OnDispose](#ondispose) `@onDispose`
+  - [OnContainerDisposed](#oncontainerdisposed) `@onContainerDisposed`
   - [Inject Property](#inject-property)
   - [Inject Method](#inject-method)
 - [Mock](#mock)
@@ -104,7 +104,7 @@ bundlers tree-shake unused exports.
 
 > [!NOTE]
 > The default `MetadataInjector` (and the `@inject` / `@onConstruct` /
-> `@onDispose` decorators) rely on `reflect-metadata`. It is declared as an
+> `@onContainerDisposed` decorators) rely on `reflect-metadata`. It is declared as an
 > optional peer dependency — install it and import it once at your entrypoint.
 > `SimpleInjector` and `ProxyInjector` do not need it.
 
@@ -181,7 +181,7 @@ Sometimes you want to get all instances from container and its scopes. For examp
 Sometimes you want to dispose a container or scope. For example, when a request, page, widget, or other local lifecycle ends.
 
 - container can be disposed
-- when container is disposed then it runs its `onDispose` hooks, unregisters its providers, removes its local instances, and detaches from its parent
+- when container is disposed then it runs its `onContainerDisposed` hooks, unregisters its providers, removes its local instances, and detaches from its parent
 
 > [!IMPORTANT]
 > Dispose is local to the container being disposed. Child scopes are not disposed automatically; dispose them explicitly when their own lifecycle ends.
@@ -415,10 +415,10 @@ Sometimes you need to invoke methods after construct or dispose of class. This i
 {{{include_file '__tests__/readme/onConstruct.spec.ts'}}}
 ```
 
-### OnDispose
+### OnContainerDisposed
 
 ```typescript
-{{{include_file '__tests__/readme/onDispose.spec.ts'}}}
+{{{include_file '__tests__/readme/onContainerDisposed.spec.ts'}}}
 ```
 
 ### Inject property
