+++ docs/src/pages/index.mdx
@@ -29,7 +29,7 @@ Awilix.
 - Clean API for registering and resolving services by class, key, token, or alias
 - No global container object; pass scopes explicitly through your app boundary
 - Tagged scopes for application, request, transaction, page, and widget lifecycles
-- Decorator support with `@register`, `@inject`, `@onConstruct`, and `@onDispose`
+- Decorator support with `@register`, `@inject`, `@onConstruct`, and `@onContainerDisposed`
 - Lazy dependencies, property injection, custom hooks, and provider decoration
 - Composable provider and registration pipes for project-specific behavior
