+++ docs/src/pages/tsyringe-alternative.mdx
@@ -31,10 +31,10 @@ objects are created.
 | ---- | -------------------- |
 | Fast TypeScript dependency resolution | Lightweight runtime with executable benchmark coverage against tsyringe |
 | Request or transaction scopes | Tagged child containers and scoped registration rules |
-| Decorator-based injection | `@register`, `@inject`, `@onConstruct`, `@onDispose`, and metadata utilities |
+| Decorator-based injection | `@register`, `@inject`, `@onConstruct`, `@onContainerDisposed`, and metadata utilities |
 | Explicit dependency tokens | Class, key, alias, function, constant, lazy, and grouped token shapes |
 | Custom factories | Provider APIs for values, classes, factories, arguments, singletons, visibility, and decoration |
-| Lifecycle cleanup | Disposable scopes and `onDispose` hooks |
+| Lifecycle cleanup | Disposable scopes and `onContainerDisposed` hooks |
 | Framework integration | Recipes for Express.js, Fastify, Next.js, React, and SolidJS |
 | Custom behavior | Pluggable injectors, provider pipes, registration pipes, and custom hooks |
