+++ lib/hooks/onContainerDisposed.ts
@@ -3,8 +3,8 @@ import type { IContainer, IContainerModule } from '../container/IContainer';
 import { HooksRunner } from './HooksRunner';
 import { type constructor } from '../utils/basic';
 
-export const onDisposeHooksRunner = new HooksRunner('onDispose');
-export const onDispose = (...fns: (HookFn | constructor<HookClass>)[]) => hook('onDispose', ...fns);
+export const onContainerDisposedHooksRunner = new HooksRunner('onContainerDisposed');
+export const onContainerDisposed = (...fns: (HookFn | constructor<HookClass>)[]) => hook('onContainerDisposed', ...fns);
 
 export type OnDisposeHook = (scope: IContainer) => void;
 
@@ -12,7 +12,7 @@ export class AddOnDisposeHookModule implements IContainerModule {
   applyTo(container: IContainer) {
     container.addOnDisposeHook((scope) => {
       for (const instance of scope.getInstances()) {
-        onDisposeHooksRunner.execute(instance, { scope });
+        onContainerDisposedHooksRunner.execute(instance, { scope });
       }
     });
   }
